Move region_scope_tree query to librustc_passes.
